+++ file/test/version_test.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: depend_on_referenced_packages
-
 @TestOn('vm')
 
 import 'dart:io';