+++ sqflite/test/utils.dart
@@ -1,0 +1,1 @@
+final fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';